extended sys args to add custom target rates + alternate recipes

+++ demo/src/main/java/Node.java
@@ -32,7 +32,7 @@ public class Node {
     }
 
     public float getModifier() {
-        int edgeSum = 0;
+        float edgeSum = 0;
         Enumeration<Float> values = edges.elements();
         while (values.hasMoreElements()) {
             edgeSum += values.nextElement();
